refactor(log analysis): extract common firmware version functionality to a single file
