+++ ardupilot_methodic_configurator/extract_param_defaults.py
@@ -14,14 +14,12 @@ SPDX-License-Identifier: GPL-3.0-or-later
 import argparse
 import contextlib
 import logging
-import re
-from typing import Any
 
 import argcomplete
 from argcomplete.completers import FilesCompleter
-from pymavlink import mavutil
 
 from ardupilot_methodic_configurator.data_model_par_dict import validate_param_name
+from ardupilot_methodic_configurator.log_analysis.backend_log_extraction import close_log, open_log
 
 NO_DEFAULT_VALUES_MESSAGE = (
     "The .bin file contained no parameter default values. Update to a newer ArduPilot firmware version."
@@ -30,12 +28,6 @@ NO_DEFAULT_VALUES_MESSAGE = (
 MAVLINK_SYSID_MAX = 2**24
 MAVLINK_COMPID_MAX = 2**8
 MAV_PARAM_TYPE_REAL32 = 9
-
-# Matches lines like "ArduCopter V4.5.5 (142aece2)", and as a fallback when, no VER message present in the log (old firmware).
-_MSG_VERSION_PATTERN = re.compile(
-    r"^(ArduCopter|ArduPlane|ArduRover|ArduSub|AntennaTracker|Blimp) "
-    r"V(\d+)\.(\d+)(?:\.(\d+))? \(([0-9a-fA-F]+)\)$"
-)
 
 
 def create_argument_parser() -> argparse.ArgumentParser:
@@ -119,109 +111,6 @@ def parse_arguments(args: list[str] | None = None) -> argparse.Namespace:
     return parsed_args
 
 
-def open_log(logfile: str) -> mavutil.mavfile:
-    """
-    Open an ArduPilot log file.
-
-    Args:
-        logfile: The path to the ArduPilot .bin log file.
-
-    Returns:
-        A mavutil.mavfile connection object.
-
-    """
-    try:
-        mlog = mavutil.mavlink_connection(logfile)
-    except Exception as e:
-        msg = f"Error opening the {logfile} logfile: {e!s}"
-        raise SystemExit(msg) from e
-    return mlog  # pyright: ignore[reportReturnType]  # pymavlink stubs include CSVReader which doesn't extend mavfile
-
-
-def close_log(mlog: mavutil.mavfile) -> None:
-    """
-    Close an ArduPilot log file.
-
-    Args:
-        mlog: The mavutil.mavfile connection to close.
-
-    """
-    with contextlib.suppress(OSError):
-        mlog.close()
-
-
-def parse_ver_fields(fws: str, maj: Any, mini: Any, pat: Any) -> tuple[str, int, int, int] | None:  # noqa: ANN401
-    """
-    Parse a VER message's core fields into (vehicle_type, major, minor, patch).
-
-    fws is the raw FWS string (e.g. "ArduCopter V4.6.3"); maj/mini/pat are the
-    raw Maj/Min/Pat field values or None.
-    """
-    fws = fws.strip()
-    if not fws:
-        return None
-    parts = fws.split(maxsplit=1)
-    vehicle_type = parts[0] if parts else ""
-    if not vehicle_type:
-        return None
-
-    if maj is None or mini is None or pat is None:
-        return None
-
-    return vehicle_type, int(maj), int(mini), int(pat)
-
-
-def _process_ver(msg: Any) -> tuple[str, int, int, int] | None:  # noqa: ANN401
-    """
-    Extract firmware version from a VER DataFlash log entry.
-
-    Returns (vehicle_type, major, minor, patch) or None if any field is missing.
-    """
-    fws = getattr(msg, "FWS", None)
-    if isinstance(fws, bytes):
-        fws = fws.decode("utf-8", errors="replace")
-    elif not isinstance(fws, str):
-        return None
-
-    return parse_ver_fields(fws, getattr(msg, "Maj", None), getattr(msg, "Min", None), getattr(msg, "Pat", None))
-
-
-def parse_msg_ver_string(message: str) -> tuple[str, int, int, int, str] | None:
-    """
-    Parse a firmware version line from MSG text, e.g. "ArduCopter V4.5.5 (142aece2)".
-
-    Also accepts a version with no patch number, e.g. "ArduCopter V4.6 (hash)", defaulting patch to 0.
-
-    Returns (vehicle_type, major, minor, patch, firmware_hash) or None.
-    """
-    matched = _MSG_VERSION_PATTERN.match(message.strip())
-    if matched is None:
-        return None
-
-    vehicle_type, major, minor, patch, firm_hash = matched.groups()
-    patch_val = int(patch) if patch is not None else 0
-    return vehicle_type, int(major), int(minor), patch_val, firm_hash
-
-
-def _parse_msg_version(msg: Any) -> tuple[str, int, int, int] | None:  # noqa: ANN401
-    """
-    Parse firmware version from a MSG DataFlash log entry.
-
-    Returns (vehicle_type, major, minor, patch) or None if the entry is not parseable.
-    The caller is responsible for not calling this once a result has already been found.
-    """
-    message = getattr(msg, "Message", "")
-    if isinstance(message, bytes):
-        message = message.decode("utf-8", errors="replace")
-    elif not isinstance(message, str):
-        return None
-    parsed = parse_msg_ver_string(message)
-    if parsed is None:
-        return None
-    vehicle_type, major, minor, patch, _firm_hash = parsed
-    return vehicle_type, major, minor, patch
-
-
 def _get_parm_value(m: object, pname: str, param_type: str) -> float | None:
     """
     Extract the float value from a PARM message for the requested param_type.
@@ -303,44 +192,6 @@ def extract_parameter_values(logfile: str, param_type: str = "defaults") -> dict
     if not values:
         raise SystemExit(NO_DEFAULT_VALUES_MESSAGE)
     return values
-
-
-def extract_firmware_version_and_vehicle_type(logfile: str) -> tuple[str, int, int, int]:
-    """
-    Extract vehicle type and firmware version from an ArduPilot .bin log file.
-
-    Prefers the structured VER message (fields Maj, Min, Pat, FWS) and falls back to
-    scanning MSG messages until one with a parseable "Vx.y" version is found
-    (e.g. "ArduCopter V4.6.3 (hash)").
-
-    Args:
-        logfile: The path to the ArduPilot .bin log file.
-
-    Returns:
-        A tuple of (vehicle_type, major, minor, patch), e.g. ("ArduCopter", 4, 6, 3).
-
-    """
-    mlog = open_log(logfile)
-    try:
-        msg_fallback_result: tuple[str, int, int, int] | None = None
-        while True:
-            m = mlog.recv_match(type=["VER", "MSG"])
-            if m is None:
-                break
-            if m.get_type() == "VER":
-                result = _process_ver(m)
-                if result is not None:
-                    return result
-            elif m.get_type() == "MSG" and msg_fallback_result is None:
-                msg_fallback_result = _parse_msg_version(m)
-
-        if msg_fallback_result is not None:
-            return msg_fallback_result
-
-        msg = f"No firmware version information found in {logfile}"
-        raise SystemExit(msg)
-    finally:
-        close_log(mlog)
 
 
 def missionplanner_sort(item: str) -> tuple[str, ...]:
