+++ tests/test_extract_param_defaults.py
@@ -12,7 +12,6 @@ SPDX-License-Identifier: GPL-3.0-or-later
 
 import unittest
 from collections.abc import Generator
-from pathlib import Path
 from unittest.mock import MagicMock, Mock, call, patch
 
 import pytest
@@ -22,7 +21,6 @@ from ardupilot_methodic_configurator.extract_param_defaults import (
     MAVLINK_SYSID_MAX,
     NO_DEFAULT_VALUES_MESSAGE,
     create_argument_parser,
-    extract_firmware_version_and_vehicle_type,
     extract_parameter_values,
     mavproxy_sort,
     missionplanner_sort,
@@ -30,8 +28,6 @@ from ardupilot_methodic_configurator.extract_param_defaults import (
     parse_arguments,
     sort_params,
 )
-
-FIXTURE_LOG = Path(__file__).resolve().parent / "fixtures" / "backend_log_80k.bin"
 
 
 @pytest.fixture
@@ -86,19 +82,16 @@ class TestArgParseParameters(unittest.TestCase):  # pylint: disable=missing-clas
 
 
 class TestExtractParameterDefaultValues(unittest.TestCase):  # pylint: disable=missing-class-docstring
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_logfile_does_not_exist(self, mock_mavlink_connection) -> None:
         # Mock the mavlink connection to raise an exception
-        mock_mavlink_connection.side_effect = Exception("Test exception")
+        mock_mavlink_connection.side_effect = OSError("Test exception")
 
         # Call the function with a dummy logfile path
-        with pytest.raises(SystemExit) as cm:
+        with pytest.raises(OSError, match=r"^Error opening logfile dummy\.bin: Test exception$"):
             extract_parameter_values("dummy.bin")
 
-        # Check the error message
-        assert str(cm.value) == "Error opening the dummy.bin logfile: Test exception"
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_extract_parameter_default_values(self, mock_mavlink_connection) -> None:
         # Mock the mavlink connection and the messages it returns
         mock_mlog = MagicMock()
@@ -115,7 +108,7 @@ class TestExtractParameterDefaultValues(unittest.TestCase):  # pylint: disable=m
         # Check if the defaults dictionary contains the correct parameters and values
         assert defaults == {"PARAM1": 1.1, "PARAM2": 2.0}
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_no_parameters(self, mock_mavlink_connection) -> None:
         # Mock the mavlink connection to return no parameter messages
         mock_mlog = MagicMock()
@@ -127,7 +120,7 @@ class TestExtractParameterDefaultValues(unittest.TestCase):  # pylint: disable=m
             extract_parameter_values("dummy.bin")
         assert str(cm.value) == NO_DEFAULT_VALUES_MESSAGE
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_no_parameter_defaults(self, mock_mavlink_connection) -> None:
         # Mock the mavlink connection to simulate no parameter default values in the .bin file
         mock_mlog = MagicMock()
@@ -139,7 +132,7 @@ class TestExtractParameterDefaultValues(unittest.TestCase):  # pylint: disable=m
             extract_parameter_values("dummy.bin")
         assert str(cm.value) == NO_DEFAULT_VALUES_MESSAGE
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_invalid_parameter_name(self, mock_mavlink_connection) -> None:
         # Mock the mavlink connection to simulate an invalid parameter name
         mock_mlog = MagicMock()
@@ -150,7 +143,7 @@ class TestExtractParameterDefaultValues(unittest.TestCase):  # pylint: disable=m
         with pytest.raises(SystemExit):
             extract_parameter_values("dummy.bin")
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_long_parameter_name(self, mock_mavlink_connection) -> None:
         # Mock the mavlink connection to simulate a too long parameter name
         mock_mlog = MagicMock()
@@ -161,7 +154,7 @@ class TestExtractParameterDefaultValues(unittest.TestCase):  # pylint: disable=m
         with pytest.raises(SystemExit):
             extract_parameter_values("dummy.bin")
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_extract_values_conversion_error(self, mock_mavlink_connection) -> None:
         """Test error handling when Value can't be converted to float."""
         mock_mlog = MagicMock()
@@ -174,7 +167,7 @@ class TestExtractParameterDefaultValues(unittest.TestCase):  # pylint: disable=m
             extract_parameter_values("dummy.bin", "values")
         assert "Error converting not_a_number to float" in str(excinfo.value)
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_extract_non_default_values_conversion_error(self, mock_mavlink_connection) -> None:
         """Test error handling when Value can't be converted to float in non_default_values mode."""
         mock_mlog = MagicMock()
@@ -187,7 +180,7 @@ class TestExtractParameterDefaultValues(unittest.TestCase):  # pylint: disable=m
             extract_parameter_values("dummy.bin", "non_default_values")
         assert "Error converting not_a_number to float" in str(excinfo.value)
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_extract_values_missing_attributes(self, mock_mavlink_connection) -> None:
         """Test handling of parameters missing Value attribute in values mode."""
         mock_mlog = MagicMock()
@@ -214,7 +207,7 @@ class TestExtractParameterDefaultValues(unittest.TestCase):  # pylint: disable=m
         assert values == {"PARAM2": 2.0}
         assert "PARAM1" not in values
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_extract_non_default_values_edge_cases(self, mock_mavlink_connection) -> None:
         """Test edge cases for non_default_values extraction."""
         mock_mlog = MagicMock()
@@ -473,7 +466,7 @@ class TestCreateArgumentParser:  # pylint: disable=too-few-public-methods
         assert args.bin_file == "dummy.bin"
 
 
-class TestSortParams:  # pylint: disable=too-few-public-methods
+class TestSortParams:
     """Tests for the sort_params function."""
 
     def test_sort_params_none(self) -> None:
@@ -482,240 +475,6 @@ class TestSortParams:  # pylint: disable=too-few-public-methods
 
         # With sort_type "none", the order should be preserved
         assert list(sorted_params.keys()) == ["B_PARAM", "A_PARAM", "C_PARAM"]
-
-
-class TestExtractFirmwareVersionAndVehicleType:
-    """BDD tests for extract_firmware_version_and_vehicle_type."""
-
-    def _make_mlog(self, messages: list) -> MagicMock:
-        """Build a mock mavlink connection that yields the given messages then None."""
-        mlog = MagicMock()
-        mlog.recv_match.side_effect = [*messages, None]
-        return mlog
-
-    def _make_ver_msg(self, fws: str, maj: int, min_: int, pat: int) -> MagicMock:
-        msg = MagicMock()
-        msg.get_type.return_value = "VER"
-        msg.FWS = fws
-        msg.Maj = maj
-        msg.Min = min_
-        msg.Pat = pat
-        return msg
-
-    def _make_msg_msg(self, text: str) -> MagicMock:
-        msg = MagicMock()
-        msg.get_type.return_value = "MSG"
-        msg.Message = text
-        return msg
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_ver_message_returns_full_four_tuple(self, mock_conn: MagicMock) -> None:
-        """
-        VER message produces a complete (vehicle_type, major, minor, patch) 4-tuple.
-
-        GIVEN: A .bin log whose first relevant message is a VER record for ArduCopter 4.6.3
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: Returns ("ArduCopter", 4, 6, 3) with all four components correct
-        """
-        ver = self._make_ver_msg("ArduCopter V4.6.3 (3fc7011a)", 4, 6, 3)
-        mock_conn.return_value = self._make_mlog([ver])
-
-        result = extract_firmware_version_and_vehicle_type("flight.bin")
-
-        assert result == ("ArduCopter", 4, 6, 3)
-        vehicle_type, major, minor, patchv = result
-        assert vehicle_type == "ArduCopter"
-        assert major == 4
-        assert minor == 6
-        assert patchv == 3
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_ver_message_extracts_correct_vehicle_type_from_fws_field(self, mock_conn: MagicMock) -> None:
-        """
-        Vehicle type comes from the first word of the FWS field, not from a separate field.
-
-        GIVEN: A VER message where FWS is "ArduPlane V4.5.1 (abcdef01)"
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: vehicle_type is "ArduPlane" and version components are (4, 5, 1)
-        """
-        ver = self._make_ver_msg("ArduPlane V4.5.1 (abcdef01)", 4, 5, 1)
-        mock_conn.return_value = self._make_mlog([ver])
-
-        vehicle_type, major, minor, patchv = extract_firmware_version_and_vehicle_type("log.bin")
-
-        assert vehicle_type == "ArduPlane"
-        assert major == 4
-        assert minor == 5
-        assert patchv == 1
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_ver_message_with_patch_zero(self, mock_conn: MagicMock) -> None:
-        """
-        VER message with Pat=0 correctly returns patch=0 rather than a default or absent value.
-
-        GIVEN: A VER message where Pat field is 0
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: The returned patch component is exactly 0
-        """
-        ver = self._make_ver_msg("Rover V4.4.0 (deadbeef)", 4, 4, 0)
-        mock_conn.return_value = self._make_mlog([ver])
-
-        vehicle_type, major, minor, patchv = extract_firmware_version_and_vehicle_type("log.bin")
-
-        assert vehicle_type == "Rover"
-        assert major == 4
-        assert minor == 4
-        assert patchv == 0
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_ver_message_takes_priority_over_msg_message(self, mock_conn: MagicMock) -> None:
-        """
-        VER message is preferred and MSG fallback is ignored when VER is present.
-
-        GIVEN: A log containing a MSG message followed by a VER message
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: Data comes from VER, not MSG; the MSG data is completely ignored
-        """
-        msg_first = self._make_msg_msg("ArduPlane V3.9.9 (stale-hash)")
-        ver_second = self._make_ver_msg("ArduCopter V4.6.3 (correct)", 4, 6, 3)
-        mock_conn.return_value = self._make_mlog([msg_first, ver_second])
-
-        vehicle_type, major, minor, patchv = extract_firmware_version_and_vehicle_type("log.bin")
-
-        assert vehicle_type == "ArduCopter"
-        assert major == 4
-        assert minor == 6
-        assert patchv == 3
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_msg_fallback_used_when_no_ver_message(self, mock_conn: MagicMock) -> None:
-        """
-        MSG message is used as fallback when VER message is absent.
-
-        GIVEN: A .bin log that contains only a MSG record for ArduCopter V4.6.3
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: Returns ("ArduCopter", 4, 6, 3) parsed from the MSG text
-        """
-        msg = self._make_msg_msg("ArduCopter V4.6.3 (3fc7011a)")
-        mock_conn.return_value = self._make_mlog([msg])
-
-        vehicle_type, major, minor, patchv = extract_firmware_version_and_vehicle_type("log.bin")
-
-        assert vehicle_type == "ArduCopter"
-        assert major == 4
-        assert minor == 6
-        assert patchv == 3
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_msg_fallback_without_patch_defaults_to_zero(self, mock_conn: MagicMock) -> None:
-        """
-        MSG fallback with only major.minor (no patch) returns patch=0.
-
-        GIVEN: A MSG record that reads "ArduCopter V4.6 (hash)" (no patch component)
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: Returns ("ArduCopter", 4, 6, 0) with patch defaulted to 0
-        """
-        msg = self._make_msg_msg("ArduCopter V4.6 (3fc7011a)")
-        mock_conn.return_value = self._make_mlog([msg])
-
-        vehicle_type, major, minor, patchv = extract_firmware_version_and_vehicle_type("log.bin")
-
-        assert vehicle_type == "ArduCopter"
-        assert major == 4
-        assert minor == 6
-        assert patchv == 0
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_first_parseable_msg_message_is_used_as_fallback(self, mock_conn: MagicMock) -> None:
-        """
-        The first MSG record with a parseable Vx.y version is used; unparsable ones are skipped.
-
-        GIVEN: Two MSG messages - first is a boot log line without a version, second is ArduCopter 4.6.3
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: Data is from the second (parseable) MSG message
-        """
-        msg1 = self._make_msg_msg("Boot started")
-        msg2 = self._make_msg_msg("ArduCopter V4.6.3 (abcdef12)")
-        mock_conn.return_value = self._make_mlog([msg1, msg2])
-
-        vehicle_type, _major, _minor, patchv = extract_firmware_version_and_vehicle_type("log.bin")
-
-        assert vehicle_type == "ArduCopter"
-        assert patchv == 3
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_raises_system_exit_when_logfile_cannot_be_opened(self, mock_conn: MagicMock) -> None:
-        """
-        A SystemExit is raised with an informative message when the logfile cannot be opened.
-
-        GIVEN: A logfile path that causes mavlink_connection to raise an exception
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: SystemExit is raised with a message mentioning the logfile path
-        """
-        mock_conn.side_effect = Exception("file not found")
-
-        with pytest.raises(SystemExit) as exc_info:
-            extract_firmware_version_and_vehicle_type("missing.bin")
-
-        assert "missing.bin" in str(exc_info.value)
-        assert "Error opening" in str(exc_info.value)
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_raises_system_exit_when_no_version_information_found(self, mock_conn: MagicMock) -> None:
-        """
-        A SystemExit is raised when neither VER nor parseable MSG is found.
-
-        GIVEN: A .bin log that contains no VER or MSG messages
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: SystemExit is raised with a message mentioning the logfile and "No firmware version"
-        """
-        mock_conn.return_value = self._make_mlog([])  # No messages at all
-
-        with pytest.raises(SystemExit) as exc_info:
-            extract_firmware_version_and_vehicle_type("empty.bin")
-
-        assert "empty.bin" in str(exc_info.value)
-        assert "No firmware version" in str(exc_info.value)
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_raises_system_exit_when_msg_version_format_is_unparseable(self, mock_conn: MagicMock) -> None:
-        """
-        A SystemExit is raised when the MSG version string cannot be parsed.
-
-        GIVEN: A MSG record whose text contains no recognisable "Vx.y" pattern
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: SystemExit is raised
-        """
-        msg = self._make_msg_msg("Boot started")
-        mock_conn.return_value = self._make_mlog([msg])
-
-        with pytest.raises(SystemExit) as exc_info:
-            extract_firmware_version_and_vehicle_type("bad.bin")
-
-        assert "bad.bin" in str(exc_info.value)
-        assert "No firmware version" in str(exc_info.value)
-
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
-    def test_return_type_is_tuple_of_str_int_int_int(self, mock_conn: MagicMock) -> None:
-        """
-        The return value has the correct types for all four components.
-
-        GIVEN: A valid VER message
-        WHEN: extract_firmware_version_and_vehicle_type is called
-        THEN: Returns a 4-tuple of (str, int, int, int) — not floats or strings for version numbers
-        """
-        ver = self._make_ver_msg("Heli V4.5.2 (cafebabe)", 4, 5, 2)
-        mock_conn.return_value = self._make_mlog([ver])
-
-        result = extract_firmware_version_and_vehicle_type("log.bin")
-
-        assert isinstance(result, tuple)
-        assert len(result) == 4
-        vehicle_type, major, minor, patchv = result
-        assert isinstance(vehicle_type, str)
-        assert isinstance(major, int)
-        assert isinstance(minor, int)
-        assert isinstance(patchv, int)
 
     def test_sort_params_qgcs(self) -> None:
         params = {"ZZZ_PARAM": 3.0, "AAA_PARAM": 1.0, "MMM_PARAM": 2.0}
@@ -739,19 +498,10 @@ class TestExtractFirmwareVersionAndVehicleType:
         assert list(sorted_params.keys()) == ["A_PARAM", "B_PARAM", "C_PARAM"]
 
 
-@pytest.mark.integration
-class TestExtractFirmwareVersionAndVehicleTypeIntegration:  # pylint: disable=too-few-public-methods
-    """Integration tests for extract_firmware_version_and_vehicle_type."""
-
-    def test_extracts_version_from_real_log_fixture(self) -> None:
-        """GIVEN a real log fixture, WHEN firmware is extracted, THEN the expected tuple is returned."""
-        assert extract_firmware_version_and_vehicle_type(str(FIXTURE_LOG)) == ("ArduCopter", 4, 6, 2)
-
-
 class TestExtractParameterValues:
     """Tests for the extract_parameter_values function."""
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_extract_parameter_values(self, mock_mavlink_connection) -> None:
         mock_mlog = MagicMock()
         mock_mavlink_connection.return_value = mock_mlog
@@ -766,7 +516,7 @@ class TestExtractParameterValues:
         values = extract_parameter_values("dummy.bin", "values")
         assert values == {"PARAM1": 1.0, "PARAM2": 2.0}
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_extract_non_default_values(self, mock_mavlink_connection) -> None:
         mock_mlog = MagicMock()
         mock_mavlink_connection.return_value = mock_mlog
@@ -783,7 +533,7 @@ class TestExtractParameterValues:
         assert values == {"PARAM1": 1.0, "PARAM3": 3.0}
         assert "PARAM2" not in values
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_invalid_param_type(self, mock_mavlink_connection) -> None:
         mock_mlog = MagicMock()
         mock_mavlink_connection.return_value = mock_mlog
@@ -797,7 +547,7 @@ class TestExtractParameterValues:
             extract_parameter_values("dummy.bin", "invalid_type")
         assert "Invalid type" in str(excinfo.value)
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_duplicate_parameter_name(self, mock_mavlink_connection) -> None:
         mock_mlog = MagicMock()
         mock_mavlink_connection.return_value = mock_mlog
@@ -812,7 +562,7 @@ class TestExtractParameterValues:
         values = extract_parameter_values("dummy.bin", "defaults")
         assert values == {"PARAM1": 1.0, "PARAM2": 2.0}
 
-    @patch("ardupilot_methodic_configurator.extract_param_defaults.mavutil.mavlink_connection")
+    @patch("ardupilot_methodic_configurator.log_analysis.backend_log_extraction.mavutil.mavlink_connection")
     def test_non_float_conversion_error(self, mock_mavlink_connection) -> None:
         mock_mlog = MagicMock()
         mock_mavlink_connection.return_value = mock_mlog
